backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
requests==2.31.0
beautifulsoup4==4.12.3
langchain==0.1.6
langchain-community==0.0.20
chromadb==0.4.22
sentence-transformers==2.3.1
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.1
google-generativeai==0.4.0
numpy==1.26.3
